refactor current_user logic into its own class and out of application_controller
